Refactor API package structure to modular monolith
Move controllers and config to feature-based modules (health, wiremock, config).
Update AGENTS.md to mandate modular monolith architecture.
Verify build and tests pass.
