+++ api/src/main/kotlin/com/example/demo/config/SecurityConfig.kt
@@ -1,4 +1,4 @@
-package com.example.demo
+package com.example.demo.config
 
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
