+++ api/src/main/kotlin/com/example/demo/health/controller/HealthController.kt
@@ -1,4 +1,4 @@
-package com.example.demo
+package com.example.demo.health.controller
 
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.RestController
