+++ api/src/main/kotlin/com/example/demo/wiremock/controller/WireMockController.kt
@@ -1,4 +1,4 @@
-package com.example.demo
+package com.example.demo.wiremock.controller
 
 import org.springframework.beans.factory.annotation.Value
 import org.springframework.web.bind.annotation.GetMapping
